+++ app/src/main/java/tn/amin/mpro2/features/util/theme/KeepThemeColorSupplier.java
@@ -2,7 +2,7 @@ package tn.amin.mpro2.features.util.theme;
 
 public class KeepThemeColorSupplier extends ThemeColorSupplier {
     @Override
-    public Long getColor(ColorType colorType) {
+    public Integer getColor(ColorType colorType) {
         return null;
     }
 }
