+++ app/src/main/java/tn/amin/mpro2/features/util/theme/ThemeColorSupplier.java
@@ -1,5 +1,5 @@
 package tn.amin.mpro2.features.util.theme;
 
 abstract public class ThemeColorSupplier {
-    abstract public Long getColor(ColorType colorType);
+    abstract public Integer getColor(ColorType colorType);
 }
